Create Webscraping,regex,write to excel
Code to extract information on number of shares issues,prices, voting rights and shares in treasury.Uses web-scraping and regex  functions to get data and writes it to a csv file. Data was saved as text on website so had o do a bit of text manipulation to get the numbers.